Implement thread architecture. Added neovim mainloop
In this commit I've created a clear line between GDB thread and nvim
thread.
Whenever we receive something from a GDB we have to make a half sync /
half async call into neovim mainloop thread.
All the stuff related to NVIM have to be done in mainloop thread as it
uses a greenlet which is ONLY valid within the mainloop thread.

TODO: Sometimes we will need to make a deep copy of GDB data before
switching thread as it's available only on the stack of that thread.
There's also a chance to implement half-sync/async both ways.